Made the instructions more explicit

+++ feedback/models.py
@@ -26,7 +26,7 @@ def validate_fourth(value):
 
 class Feedback(models.Model):
     field1 = models.CharField(max_length = 200, validators =[validate_first] )
-    field2 = models.PositiveSmallIntegerField(validators=[MinValueValidator(5340), MaxValueValidator(5350)])
+    field2 = models.PositiveIntegerField(validators=[MinValueValidator(5340), MaxValueValidator(5350)])
     field3 = models.CharField(max_length = 200, validators =[validate_third] )
     field4 = models.CharField(max_length = 200, validators =[validate_fourth]  )
     name = models.CharField(max_length = 200)
